progressing with other names screen
