+++ src/main/steps/applicant1/other-names/post.ts
@@ -30,6 +30,7 @@ export default class FindAddressPostController extends PostController<AnyObject>
           req.session.userCase.additionalName = '';
         }
       }
+      console.log(formData);
     }
 
     const nextUrl =
