--- src/main/steps/common/models/summaryListRow.ts
@@ -1,11 +1,0 @@
-export type ActionItem = {
-  href?: string;
-  text: string;
-  visuallyHiddenText?: string;
-};
-
-export type SummaryListRow = {
-  key?: string;
-  value: string;
-  actions: ActionItem[];
-};
